Shorted shared memory names to avoid error on macOS
macOS shared memory names are limited to 31 characters, and using the
full timestamp could sometimes push the name of the shared memory over
that limit. To resolve this, this commit changes shared memory names to
use the time as an integer rounded to the nearest millisecond rather
than the full timestamp.